大量更新

+++ src/Library/Data/Core/Data.Abstractions/SqlQueryable/GroupByQueryable/INetSqlGrouping.cs
@@ -7,6 +7,6 @@
     {
         TKey Key { get; }
 
-        int Count();
+        long Count();
     }
 }
